+++ src/Cubes/momentMap.cc
@@ -109,13 +109,13 @@ namespace duchamp
     std::vector<bool> objectChoice = this->par.getObjectChoices(this->objectList->size());
 
     bool *isObj = new bool[xdim*ydim*zdim];
-    for(int i=0;i<xdim*ydim*zdim;i++) isObj[i] = false;
+    for(size_t i=0;i<xdim*ydim*zdim;i++) isObj[i] = false;
     for(size_t i=0;i<this->objectList->size();i++){
       if(objectChoice[i]){
 	std::vector<Voxel> voxlist = this->objectList->at(i).getPixelSet();
 	std::vector<Voxel>::iterator vox;
 	for(vox=voxlist.begin();vox<voxlist.end();vox++){
-	  int pixelpos = vox->getX() + xdim*vox->getY() + xdim*ydim*vox->getZ();
+	  size_t pixelpos = vox->getX() + xdim*vox->getY() + xdim*ydim*vox->getZ();
 	  isObj[pixelpos] = true;
 	}
       }
@@ -150,7 +150,7 @@ namespace duchamp
 	double * world = this->head.pixToVel(x,y,zArray,zdim);
       
 	for(int z=0; z<zdim; z++){      
-	  int pos =  z*xdim*ydim + pix;  // the voxel in the cube
+	  size_t pos =  z*xdim*ydim + pix;  // the voxel in the cube
 	  if(isObj[pos]){ // if it's an object pixel...
 	    // delta-vel is half the distance between adjacent channels.
 	    // if at end, then just use 0-1 or (zdim-1)-(zdim-2) distance
